LPS-61156 Rename
